Use long instead of intptr_t for IDs

+++ Source/JavaScriptCore/inspector/JavaScriptCallFrame.h
@@ -41,7 +41,7 @@ public:
     }
 
     JavaScriptCallFrame* caller();
-    size_t sourceID() const { return m_debuggerCallFrame->sourceID(); }
+    long sourceID() const { return m_debuggerCallFrame->sourceID(); }
     const TextPosition position() const { return m_debuggerCallFrame->position(); }
     int line() const { return m_debuggerCallFrame->line(); }
     int column() const { return m_debuggerCallFrame->column(); }
